Ugly fix for embedded commas; reformat comments; add nascent SPOT utilities
Some of the EFO trait labels in the GWAS spreadsheet contain embedded commas,
which causes the trait array to be too long (broken into too many strings).
Implement fix which checks whether the EFO URI and trait arrays are the same
length: if the latter is longer, search for the shortest string apart from the
first string in the list. Append this shortest string to the preceding string in
the list and then delete the shortest string from the list as a standalone
entry. The list of traits consequently shortens by one. Iterate until the URI
and trait arrays are the same size. It is not pretty, but it is a solution which
works in most cases.

We try to reserve triple-quoted comments for actual help text, whereas commented
code should be preceded with hash marks.

Add single-function module 'spotilities', which provides for printing to
standard error; this fixes some problems with output redirection and piping. All
error and informational messsages related to processing should be sent to the
error console, really.